tests/functional_tests/requirements.txt
numpy==1.22.4
pytest-timeout==2.2.0
python-dotenv==0.21.0
scikit-learn==1.1.2
tornado==6.0.4
